CMake goodies: - adding moving helper macros to cmake/Helpers.cmake, and revamping them as functions, capitalizing the function names as well.
Dealing with compiler flags in CMake.  Before we had a separate variable Lass_CXX_FLAG we used to store extra required flags.
This was convenient because we didn't need to FORCE anything in CMAKE_CXX_FLAGS, but it's inconvenient as it circumvents
the user preferred way to add/remove flags: CMAKE_CXX_FLAGS. 
To solve, we now add the flags directly to CMAKE_CXX_FLAGS, but we try to keep track of what we're doing:

- adding LASS_ADD_FLAGS_ONCE to set flags once in cached variables like CMAKE_CXX_FLAGS, and not set it again if it was removed.
- adding LASS_LINK_OPTION_TO_FLAG to link an OPTION variable like BUILD_WITHOUT_ITERATOR_DEBUGGING to a flag in CMAKE_CXX_FLAGS
- all additional compiler flags are now stored in CMAKE_CXX_FLAGS, instead of a separate variable Lass_CXX_FLAG
- this makes it easier to pick up compiler flags that needs to be exported, using LASS_GATHER_EXPORTED_FLAGS

git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1213 e9412239-ebb7-49aa-8f4a-f9aa784616d7