Conan: python_version and python_executable options
The python option was not sufficient to fully determine the Python
version when Python's path is injected as some build requirement
(via dep_env_info): package_id would run without the updated PATH of
dep_env_info, and the wrong Python would be found to determine the
version. To circumvent this, we allow to explicitly set python_version
as option.

In build and package, when then CMake project is configures, PATH will
be updated with the dep_env_info, so the correct Python is found and the
configured python_version will verify that.