Conan: fix self.options.with_std_auto_ptr check.
with_std_auto_ptr has three valid settings: True, False, or not set.
Not set is represented by None, but apparantly, by the time you want to
test it, its value has become the _string_ "None".

It's not clear from the documentation of Conan if this is a bug or
intended behaviour. In any case, we're testing both.