Use std::ptrdiff_t instead of ptrdiff_t
This is more consistent.  There was a mixture of std::ptrdiff_t and
ptrdiff_t in different files. We're now using std::ptrdiff_t everywhere.
This also means we no longer need to #include <stddef.h> but <cstddef>.
AFAIK, we can always assume it to be available, so no need to test it in
Environment.cmake