Conan: use python executable to determine link library.
This mostly matters to Windows builds of Lass. On Linux this should not
be an issue.

If you want to explictly link to a release build of Python on Windows,
the only surefire way is to set both Python_LIBRARY_RELEASE and
Python_LIBRARY_DEBUG to the release library, before running cmake.

FindPython.cmake and FindPythonLibs.cmake are very good at picking up a
debug library via the registry. Trying to avoid one by setting PATH (*)
to a python build that only contains the release library, simply doesn't
work. If a debug library of the same version can be find in an installed
Python distro via the registry, it _will_ be found.

So by explicitly setting Python_LIBRARY_DEBUG to the same release
library as Python_LIBRARY_RELEASE, you can force it.

(*) PATH can be set via the profile, either directly, but also as a
build_requires. Care must be taken to not query the python executable
outside of the _cmake method. In particular, you should not query
python in configure or package_id, because PATH will not be fully formed
yet. If you're using build_requires to influence PATH, this PATH will
_not_ be seen until the build method. configure and package_id would
simply use the wrong python executable.
For this to work, you must explicitly specify the python_version and
python_debug options, because if they are not set, then python _must_ be
evaluated in configure.