Fix Python 3 implementation ShadowTraits<Map>
In Python 3, dct.items() does not return a list, so you need to again
use PyMapping_Items to get the list of items. In Python 2, we directly
called PyObject_CallMethod(obj, (char*)"items", 0) because of a
const char* to char* casting warning, and it was also exactly that what
PyMapping_Items was doing. But In Python 3, it's a more elaborate
function.