Replace TypeError in python::Map __delitem__
When deleting an item in the dictionary, and the key cannot be found, a
KeyError is raised. But when the key could not be cast to the key_type
of the underlying container, a TypeError was raised.

This was inconsistent with with __getitem__ and __contains__: if the key
cannot be cast to the key_type, it's certainly not in the
dictionary. And thus instead of raising a TypeError, a KeyError should
be raised, or False should be returned.

This is now also true for __delitem__: if the key cannot be cast to the
key_type, a KeyError is raised instead of a TypeError.