Fix error propagation in objobjargproc slots
The only slot like this is mp_ass_subscript and is used to implement
both __setitem__ and __delitem__ (the latter gets a NULL as value)

mp_ass_subscript has an integer return code, 0 on success and -1 as
error indicator. Instead of simply wrapping the integer into a PyObject, it should be
correctly propagated using a NULL pointer on error. On success, we
simply return None.