Add testcase for dual const/non-const getter
PY_CLASS_MEMBER_RW and PY_CLASS_MEMBER_R need at least one const getter,
but some types also have a non-const getter that may or may not have the
same return type.

A relevant use case is when the const version returns a pointer to a
const object, and the non-const version returns a pointer to a non-const
object. In the latter case you will be able to edit the member in
Python, but not in the former case.