Undefine slots for compatibility with Qt
When using Lass (with Python>=3.2) in a project
with Qt, the macro `slots` may be defined for usage in Qt. `slots` is
also used as a member name of `PyType_Spec` in `object.h`. It is
sufficient to temporarily undefine `slots` since this member is not
being used on the client side.

`push_macro` is supported by MSVC, GCC and clang.