Conan: explicit error if python cannot be found.
If python_executable cannot be found, python.executable will be None.
Instead of letting this result in an obscure TypeError: expected str,
bytes or os.PathLike object, not NoneType; raise an explicit
ConanInvalidConfiguration stating the actual cause.

One situation in which this can happen is on a Linux system with only
python3 installed, and the default python executable can indeed not be
found.