Fix deprecated-copy warnings
As of GCC 9, a compiler warning can be emitted if a class has an
implicitly declared copy constructor or copy assignment operator, while
that class already has an user-defined copy assignment operator or copy
constructor. This is because in C++11, this has indeed been deprecated.
This warning is enabled by -Wdeprecated-copy, also included in -Wextra.

As a mitigation, a user-defined copy constructor or copy assignment
operator has been added.

In case of the various Callback classes, there was an templated
conversion assignment operator that could act as copy assignment
operator, but was not recognized as such. However, this conversion
assignment operator is not required. With a normal copy assignment
operator, the constructor can do the conversion _before_ assignment,
instead of doing so _inside_ the assignment operator.