Fix CMP0075 policy warning
CMake 3.12 changed the way check_include_file behaves regarding to
CMAKE_REQUIRED_LIBRARIES. This triggered a policy warning because
CMAKE_REQUIRED_LIBRARIES was still set to pthread after using it for
detecting pthread_setaffinity_np. By unsetting this afterwards (which is
a good idea anyway), we avoid the situation in which this policy
matters.

  Policy CMP0075 is not set: Include file check macros honor
  CMAKE_REQUIRED_LIBRARIES.  Run "cmake --help-policy CMP0075" for
  policy details.  Use the cmake_policy command to set the policy and
  suppress this warning.
  CMAKE_REQUIRED_LIBRARIES is set to:
    pthread
  For compatibility with CMake 3.11 and below this check is ignoring it.