Always use Python_LIBRARY_{RELEASE|DEBUG}
In FindPythonCompat, if CMake < 3.12, the older FindPythonLibs is used
that relies on PYTHON_LIBRARY and PYTHON_DEBUG_LIBRARY to specify the
Python libraries.

If they're not initialized, but Python_LIBRARY_{RELEASE|DEBUG} are, then
set them accordingly, so we can always use
Python_LIBRARY_{RELEASE|DEBUG} in conanfile to initialize them.