Add Python::Module to FindPythonCompat
This was introduced in FindPython in CMake 3.15

When building a Python module, you should not be linking to the Python
library Python::Python. The symbols are already provided for by the
Python::Interpreter.  You only need the include directories.  These are
provided by Python::Module: it replaces Python::Python as it only provides
the include directories.

Only when embedding a Python interpreter, you should link to
Python::Python. For example, when building the test suite.

So the targets lass_python and pylass should only link to
Python::Module. If you want use it with an embedded interpreter, it's
your duty to link to Python::Python as well.

On Windows, Python::Module is the same as Python::Python.

This was discovered when trying to build pylass on manylinux2010-x64:
libpython* was not available, as it turns out this is not needed when
building wheels.