Fix missing-field-initializers warning, Python 3.8
In Python 3.8, PyTypeObject is being modified to support Vectorcall
functions: tp_vectorcall is added and tp_print recycled as to
tp_vectorcall_offset.

The fields exists in Python 3.8, but it is only used since 3.9.

tp_print is moved to the bottom of the structure as a deprecated field
that only exists in Python 3.8. So we must initialize it to avoid the
missing-field-initializers warning, but then we get a deprecation
warning which we now suppress.

(cherry picked from commit 3726ee379d59b56bd8d57fdadb713b62399fccb2)