Use clean versions for releases
To release a version X.Y.Z, the commit is tagged as "lass-X.Y.Z". This
is no longer a pre-release and will get a clean version X.Y.Z instead of
a pre-release version "X.Y.Z-prerelease+revision".

Post-releases are not allowed with semver, so if describe does not
return exactly "lass-X.Y.Z" (but something like "lass-X.Y.Z-N-revision")
then base_tag must be set to the current "lass-X.Y.Z" and the
CMakeLists.txt version must be incremented

We needed to add --tags to the git describe commit to allow unannotated
tags to be used.