Fix usage of _Py_ForgetReference
In Python 3.9, _Py_ForgetReference is no longer defined if Py_TRACE_REFS
is not defined.

In Python 3.8, it's still defined, but if Py_TRACE_REFS is not defined,
it only calls _Py_INC_TPFREES. That is a no-op, unless COUNT_ALLOCS is
defined. But its counterpart _Py_INC_TPALLOCS is called in
_Py_NewReference which is used in PyObjectPlus's constructors, so we
call _Py_ForgetReference in the destructor even if Py_TRACE_REFS is not
defined.

In Python 3.9, _Py_INC_TPFREES and _Py_INC_TPALLOCS no longer exist, so
they're also not part of _Py_NewReference and _Py_ForgetReference
anymore. Thus _Py_ForgetReference is a real no-op if Py_TRACE_REFS is
not defined, and they choose to no longer define _Py_ForgetReference.