Fix bug in HeaderPfm::readFrom
When reading the last value from an istream without any whitespace that
follows (reading the last integer from a line "1024 512" that does not
have an end-of-line character \n), you have reached the end-of-file and
eofbit is set in the stream's state. The consequence is that the stream
is no longer good().

But it didn't fail(). I.e. the failbit is not set, you didn't attempt to
read _past_ the end-of-file yet. So instead of testing if the stream is
still good() after reading, we should test if hasn't fail()'ed. You
should only use good() _before_ reading ...

Testing the boolean value of the stream returns the same as fail(), so
that that matches perfectly with wrapping the IO operation in an
if-statement.