Add PyExportTraits for std::filesystem::path
Accepts bytes, str or objects that implement the os.PathLike interface
(having __fspath__()). If conversion is needed from bytes to str (on
Windows) or from str to bytes (on POSIX), then it is performed using
Py_FileSystemDefaultEncoding (which is usually UTF-8, but not always).

Always converts them to str instances