Export enums as Python enum.Enum
For each exported enum, a new Python type will be defined, inheriting
from enum.Enum. This enum type must be added to a module, or another
class.

There are two possible ways to export enums:

- as integer enums, deriving from enum.IntEnum, where the Python enum
value maps directly on the C++ enum value. Because IntEnum is used, the
Python enum instances will also be valid Python ints. Python ints will
also be accepted when converting from Python to C++.

- as string enums, deriving from enum.Enum and str, where the Python
enum value is a str which needs to be defined in the enum definition.
The Python enum instances will also be valid Python str instances, and
str instances will be also be accepted when converting from Python to
C++.