Fix initialization of static atomics in Singleton
This fixes a problem with the initialization of the static std::atomics
for VS 2017, and in particular the initialization of the semaphore to
the value of 1.

If I understand correctly, static objects of std::atomic should be able
to use constant initialization [1] to initialize the semaphore to 1 at
compile-time. I.e. there should be no runtime initialization necessary
of which the initialization order would become relevant (strictly
speaking, it doesn't need to be performed at compile-time, but they must
be performed before any other initialization)

At first, one had to use ATOMIC_VAR_INIT [2] to initialize std::atomic
objects (even to default-initialize them?), and if they had static
storage duration, this would be constant initialization. That caused so
much confusion that this was later fixed to do "The Right Thing" [3] so
that normal initialization should work as expected, and ATOMIC_VAR_INIT
was deprecated.

That said, it seems neither method seems to work in VS 2017. "The Right
Thing" wouldn't be fixed until VS 2019 [4], but also the ATOMIC_VAR_INIT
method doesn't seem to help to initialize the semaphore to the value of
1: all std::atomic would be zero-initialized.

The solution is to go back to local static objects (defined in
functions). Even if no constant initialization is used, it is guaranteed
to be initialized just-in-time. Even more, since C++11, it is also
guaranteed to be thread-safe [5].

[1] https://en.cppreference.com/w/cpp/language/constant_initialization
[2] https://en.cppreference.com/w/cpp/atomic/ATOMIC_VAR_INIT
[3] https://wg21.link/P0883
[4] https://devblogs.microsoft.com/cppblog/cpp17-20-features-and-fixes-in-vs-2019/
[5] https://en.cppreference.com/w/cpp/language/storage_duration#Static_local_variables