Use variadic templates in pyobject_call.inl
Replace some of the generated functions with variadic templates.
I wasn't able to replace all of them yet, so the file still needs
to be expanded.

Most of the replacements were straight-forward, as they simply forward
function calls.

ExplicitResolver is more interesting as you need to be able to pass a
type-list as single template argument, instead of passing all types as
individual template arguments. However, using specialization, we can
easily unwrap a TypeList into its indivual types.

A type-tuple is easier two unwrap in a single step, so I introduced
TypeTuple, and replaced all usage of TypeList in pyobject_macros.h by
TypeTuple. All macros still work the same as before, and TypeList is
still accepted by the *_EX macros, but TypeTuple is now preferred.