Support tzinfo in PyExportTraits of std::chrono::time_point
When building a datetime.datetime object from a time_point, we keep
building naive datetime objects in local time, as that's the most
universally supported [1]

However, when getting a datetime object, we can do better than throwing
an error when tzinfo != Py_None: use the utcoffset to calculate UTC
time, and use the non-standard timegm or _mkgmtime to convert UTC time
to a POSIX timestamp [2]

This way we can convert all datetime objects to
system_clock::time_point, regardless of the timezone (but naive objects
are always interpreted as local time).

[1] There's nothing "localtimey" about std::chrono::system_clock. It's
just the local machine's wallclock measured in POSIX timestamps where
are always "UTC" [2]. So one could argue the (better?) alternative would
be to always produce datetime objects in UTC, with an tzinfo set to UTC.
However, C++20 introduced std::chrono::utc_clock for which we'll build
UTC datetimes, if that's you want; and wallclock is usually presented
as localtime

[2] POSIX timestamps are always referenced to 1970-1-1 midnight UTC, so
the same moment in time has the same POSIX timestamp, regardless of the
timezone you're in.