Replace volatile by std::atomic in IdGenerator
GCC 11.2.0 printed this warning:

warning: ‘++’ expression of ‘volatile’-qualified type is deprecated
[-Wvolatile]

Checking value of currentId against maximum value _before_ incrementing,
as relying on the wrapping around on (signed) integer overflow is in
fact undefined behavior.

Interestingly enough, when you add two chars together, you get an int,
which is why GCC complains when adding 1 to currentId with:

 warning: conversion to 'TValue {aka char}' from 'int' may alter its
 value [-Wconversion]

*even* when I replaced the 1 by num::NumTraits<TValue>::one!
So I've settled for a static_cast.