Add PyExportTraits for std::variant
Using the PyExportTraits of the underlying alternative types.

Converting a std::variant from C++ to Python is straightforward: we
simply use the PyExportTraits of the alternative type.

From Python to C++, that's another matter, as we need to guess the right
alternative type. The only way possible is to simply try each
alternative one by one, from left to right. And then you can still have
unexpected results:

In Python, a bool is an int is a float is a complex number: if the
complex type is put before the bool as alternative type, then any Python
bool will be accepted as a C++ complex number. You need to make sure you
order your alternative types correct: first bool, then ints, then
floats, then complex numbers.

The same is true for strings and sequences of strings: in Python, a
str is also a sequence of (one character) strings. So if one of your
alternative types is for example std::vector<std::string>, it will
happily accept single strings as value, and break them up into a vector
of single character strings. This is especially surprising if
std::string itself follows as alternative!

Booleans already get extra special treatment, as they accept anything
that can evaluate to a boolean, so you wouldn't be able to put them in
front of the alternative list. So in case of std::variant, we demand a
boolean is really a Python bool.

TODO: can we automatically sort the types? should we be more strict with
other types?

First I had PyExportTraitsVariantImpl simply as a private inner class
Impl, but GCC complained on the fully specialized version (it doesn't
need the Variant template argument):

  template <> struct Impl<> {/*...*/};

with the following error:

  error: explicit specialization in non-namespace scope

Yet, according to [1] this should be allowed in C++17 (which we surely
have to support std::variant). This appears to be a GCC bug [2]

[1] https://wg21.cmeerw.net/cwg/issue727
[2] https://stackoverflow.com/a/49707378