Force python_embedding test module in UTF-8 mode
On ubuntu jammy and kinetic (both Python 3.10), the filesystem export
tests failed because Py_FileSystemDefaultEncoding was initialized to
"ascii". The `python` executable doesn't seem to suffer from this
problem.

Since Python 3.8, we can explicitly put the embedded Python explicitly
in UTF-8 mode [1] during pre-initialization [2].

We're only doing this for the test module, not in general, until further
notice.

[1] https://peps.python.org/pep-0540/
[1] https://docs.python.org/3.8/c-api/init_config.html#preinitialization-with-pypreconfig