Fix std::u8string PyExportTraits tests
When I first added u8 literal strings to the tests, I did so on Windows
with MSVC, and I've used \x escape sequences to encode code points, i.e.
Unicode characters, as hexadecimals like \x2653, which is a funny H: ♓

However, Clang interprets them as code _units_, which in UTF-8 are a
single byte < 256, and thus we get the error:

error: hex escape sequence out of range

I don't know for sure who's right, but from [1], it seems the "byte"
interpretation of Clang is right.

So to keep Clang happy, I've replaced the escape sequences for code
points by those for code units. \x2653 became \xe2\x99, and so on.

But now MSVC would have the unit tests fail, because _it_ interpreted
\xe2\x99 again as code points instead of code units, and expanded it
again to utf-8 as \xc3\xa2\xc2\x99, which is â\u0099 (code point 0x99
isn't a real character, but a Single Graphic Character Introducer).

To keep everyone happy, we now use the \u and \U escape sequences [1],
which at least encode unicode code points on all compilers.

[1] https://en.cppreference.com/w/cpp/language/escape