Improve SimplePolygon2D area
When the polygon is far away from the origin, using the origin as third
vertex of the triangles results in loss of precision due to the
multiplication. It's better to offset the vertices so the polygon is
nearer to the origin. Any offset will do, like the first vertex of the
polygon.

https://www.johndcook.com/blog/2018/09/26/polygon-area/