Add header to nodes in AllocatorConcurrentFreeList
Before, the unallocated memory itself would be used to store the pointer
linking the nodes in the free-list. This works fine in a single-threaded
context, but it can fail in a concurrent context: it is possible that a
thread is still holding a pointer to the deallocated memory and trying
to read it (even if it will then realize it was invalid). Putting the
pointer in the deallocated memory will change its memory layout and may
cause a hang or a crash. This may happen in lock_free_queue::pop or
lock_free_stack::pop_node.

By putting the node pointer in a header in front of the allocated memory
block, the deallocated memory will not need to change its layout and
because it's not freed to the OS either, it's relatively safe for
another thread to still access it.

The size of the header will be the alignment size of the allocated
memory, so that the whole memory block retains the same alignment. This
alignment shall be at least the size of the node pointer.