Use std::atomic on TaggedPtr
We see the return of the 48|16 bit TaggedPtr due to a lack of lock-free
support on std::atomic for PODs that are wider than one pointer.

Although all 64-bit x86 processors except for the early AMD64 ones have
cmpxchg16b, it seems std::atomic does not use it. It seems for a couple
of reasons:
  1. To retain ABI compabilitity with these early AMD processors.
  2. Although lock cmpxchg16b can be used for the compare_exchange
     implementation, the other operations are not directly supported.
  3. Because of (2) performance of a lock-free 16byte std::atomic is not
     as expected.

Also align the 32-bit TaggedPtr on a 64-bit boundary. Before we
explicitly did *not* do that because msvc started complaining about ebx
being used in assembly blocks. However, as we:
  a. no longer used said assembly blocks (thanks to std::atomic),
  b. get the error message below instead,
we again can and should use proper alignment:

error C2338: You've instantiated std::atomic<T> with sizeof(T) equal to
2/4/8 and alignof(T) < sizeof(T). Before VS 2015 Update 2, this would
have misbehaved at runtime. VS 2015 Update 2 was fixed to handle this
correctly, but the fix inherently changes layout and breaks binary
compatibility. Please define _ENABLE_ATOMIC_ALIGNMENT_FIX to acknowledge
that you understand this, and that everything you're linking has been
compiled with VS 2015 Update 2 (or later).