Track shadow objects in map
Before, for each C++ object being passed on to Python (by shared
pointer), a new Python shadow object was created, even when a shadow
object for the same C++ object was still active. As result, there could
be more than two Python objects shadowing the same C++ object, and they
would have different IDs, and not evaulate as being the same object.
I.e. `x is y` would always be False if `x` and `y` were passed to Python
seperately, while shadowing the same C++ object.

In this change, each shadow object that is currently alive is tracked in
a single map. When a C++ object is being passed to Python, the map is
consulted to see if there currently exists a Python object that is
already shadowing the C++ object. If so, instead of creating a new
Python shadow object, the existing one is used (after incrementing its
reference count), and `x is y` will be True.

Limitations:

- When the last reference to a Python shadow object goes out of scope
  (its refcount goes to zero), it is obviously deallocated and removed
  from the map. That means that the next time the same C++ object is
  being passed to Python, a new Python shadow object will be created.

  In other words, we do not keep an extra reference to the shadow object
  around if it's no longer referenced from Python, just in case the same
  C++ object might be passed to Python again. Doing so would effectively
  create a memory leak, as we would have no mechanism to figure out when
  C++ object is destroyed and the shadow object is no longer needed.

  Pratically, this means that the id of a Python shadow object (`id(x)`
  should not be used _beyond_ the lifetime of the shadow object to
  identify if two Python objects shadow the same C++ object. `id(x)`
  effectively returns you the address of the Python object, meaning that
  if a new Python shadow object is created, it will likely have a
  different address and id, and that these addresses or ids can also be
  recycled for shadowing different shadow objects.

  In other words: `id(x) == id(y)` will only be true if both `x` and `y`
  are both in scope and shadow the same C++ object, even when passed to
  Python at different calls.

- This only works when passing C++ objects to Python by (shared)
  pointers. When passing C++ objects by copy, a new C++ object will made
  to be shadowed by the Python object. The next time the same original
  C++ object is passed to Python, a new C++ copy will be made and thus a
  new Python shadow object, even if the previous one is still in scope.