Fix initialization of tp_name and qualname
Say we're exporting an enum `E` on an inner class `D` on a top-level
class `C` in a module `b` in a package `a`.

The "full name" of the inner class `D` would be `a.b.C.D`, and:

- its `__name__` is `"D"`
- its `__module__` is `"a.b"`
- its `__qualname__` is `"C.D"` [1]
- but its `tp_name` needs to be "`a.b.D` (without the `C`!) because it
  initializes `__module__` and `__name__` [2] (and not `__qualname__`!)
  We need to set `__qualname__` explicitly as class attribute.

Likewise, the "full name" of the enum `E` would be `a.b.C.D.E`, and:

- its `__name__` is `"E"`
- its `__module__` is `"a.b"`
- its `__qualname__` is `"C.D.E"` [1]

`scopeName` in the C++ thus means the `__qualname__` of the outer class
(i.e. `"C"` in case of `D` and `"C.D"` in case of `E`). I avoided the
name `parentClass` in order not confuse it with class inheritance.

[1] https://peps.python.org/pep-3155/
[2] https://docs.python.org/3/c-api/typeobj.html#c.PyTypeObject.tp_name