Add bidirectional export traits for std::function
One particular bug in lass_stubgen was in parsing the parameters of
functions that return a std::function<R(T, U)> (with R, T, U actual
types like int, not template paramters): the parameters T and U of the
_return_ type would also be counted as parameters of the function.
That's because in libclang's AST the components of the function
signature gets flattend and we can't readily distinguish between
PARAM_DECL of actual parameters and template arguments:

  FUNCTION_DECL testFunctionFromCpp type=std::function<int (int, int)> (int)
    NAMESPACE_REF std type=
    TEMPLATE_REF function type=
    PARM_DECL  type=int   ← 1st parameter of the function in return type
    PARM_DECL  type=int   ← 2nd parameter of the function in return type
    PARM_DECL factor type=int <- Actual parameter of testFunctionFromCpp

In this case, only the actual parameter has a name ("factor"), but we
can't rely on this in general. The most reliable check that I've found
was that only the actual parameters have the function declaration as
semantic parent.