Do not generate lass cmakedeps
When Lass is being used as a dependency in a Conan project, Conan would
automatically generate a lass-config.cmake for the downstream project to
use. However, Lass already generates its own LassConfig.cmake.

Instruct Conan not to generate any for lass, but tell it to add the
share/Lass directory to the search path ...