Guard against unexpected line endings
When core.autocrlf=true, files in the Git repo (more technically
correct: files in the index) with LF line endings will be checkout out
with CRLF line endings on Windows. As Conan will compute a hash of the
exported source files to determine the recipe's revision ID, this means
that the same source code will generate a different revision ID on
Windows and on Linux (because on Linux those files will still be
checkout out with LF line endings). To avoid this, we must make sure
files with LF line endings are always checked out in the worktree with
LF line endings, even on Windows. [1,3]

One could rely on "asking" all developers to either set
core.autocrlf=false, globally or locally, but this is at best
error-prone. However, we can also use .gitattributes to tell how files
should be checked out for this project [2].

The only snare is that this won't automagically fix line endings in
already cloned repos. Therefore, we add a check in export_sources to
verify all line endings are expected, and print a message on how can fix
the situation.

[1] https://docs.conan.io/2.0/knowledge/faq.html#error-obtaining-different-revisions-in-linux-and-windows
[2] https://www.aleksandrhovhannisyan.com/blog/crlf-vs-lf-normalizing-line-endings-in-git/
[3] https://github.com/conan-io/conan/issues/8391