Support editable mode for Conan package
In editable mode, we only have the source and build directory (not the
packaged content). This results in a different directory layout. We can
fix most in layout() using self.cpp.source and self.cpp.build, but
loading LassConfig.py is more difficult as we need to do that in
package_info() we we can't really know if we're in editable mode or not.
So we need to hack it, and assume the build dir is named "build".