Also package PDBs of static builds
Can't use $<TARGET_PDB_FILE:tgt> for static libraries, because that
causes following error:

  CMake Error:
    Error evaluating generator expression:
      $<TARGET_PDB_FILE:lass>
    TARGET_PDB_FILE is allowed only for targets with linker created artifacts.

Indeed, TARGET_PDB_FILE [1] yields the PDB_NAME [2] target propertie,
which is a _link-time_ property.  static libraries don't invoke the
linker, so PDB_NAME is not used. What we need is the COMPILE_PDB_NAME
property [3], but there's no generator expression that yields that.

We're not setting COMPILE_PDB_NAME or COMPILE_PDB_OUTPUT_DIRECTORY, but
are relying on them being created next to the static library by default.

We're also installing them to the lib directory instead of bin.

[1] https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html#genex:TARGET_PDB_FILE
[2] https://cmake.org/cmake/help/latest/prop_tgt/PDB_NAME.html
[3] https://cmake.org/cmake/help/latest/prop_tgt/COMPILE_PDB_NAME.html