Export compile commands
This will generate a file compile_commands.json with compile commands
for all files [1,2]. This only works for Makefile or Ninja based
generators (so in particular not with Visual Studio projects).
It's not just for Clang, it works for GCC as well.

You can use this file to feed information to ccpcheck using the --project
flag [3], or use it in .vscode/c_cpp_properties.json [4]

[1] https://clangd.llvm.org/design/compile-commands
[2] https://cmake.org/cmake/help/v3.27/variable/CMAKE_EXPORT_COMPILE_COMMANDS.html
[3] http://cppcheck.net/manual.pdf#section*.17
[4] https://code.visualstudio.com/docs/cpp/c-cpp-properties-schema-reference