Fix QuadTree ray-find
It was discovered with Clang 14 (Ubuntu Jammy) that some objects would
be missed along the ray. We tracked it down to tMiddle not being robust,
having error being accumulated in it, and therefore skipping over nodes
that should have been visited.  Recomputing tMiddle from the node's
center instead of (tNear+tFar)/2 gave the true bounds of the child nodes
and gave correct results. The fact that center is computed (as the average
of the current node's min and max) is no problem, as that's exactly how
corners of the subnodes have been generated.

Then I've noticed the cause of the (tNear+tFar)/2 deviation: the fact
that the bounds of the childnodes are grown by 0.1%: the center of the
first generation child nodes will still be correct, but from the second
generation, they will shift away from the intended bisect. Removing this
artificial growth also fixed the ray-find test, but there's still floating
point error accumulating in (tNear+tFar)/2, so recomputing tMiddle from
the node's center is still more robust, and so we're keeping that.