Remove fPIC option if shared
When building shared libraries on Linux, you always need to build with
fPIC, so you can't turn off.

We remove fPIC on Windows in config_options (not just for MSVC), because
it is as if the option doesn't even exist for Windows. I.e. we remove it
before you can even _set_ the option.

But we remove fPIC when shared in configure, because it depends on another
option, so you can't do that before.
Using rm_safe in case of a double delete.

See also [1] and [2]

[1] https://docs.conan.io/2.0/tutorial/creating_packages/configure_options_settings.html
[2] https://docs.conan.io/2.0/reference/conanfile/methods/config_options.html#auto-shared-fpic