Set CMAKE_LIBRARY_OUTPUT_DIRECTORY to lib
This variable sets the location where *.so are located in the build
directory [1]. They should go in the lib directory, because that's also
the directory they are located in when installed in the package.
See `_lass_install_library`.

This variable also determines where MODULE libraries are located, in
particular lass.so/lass.pyd. On Linux, lass.so is instaleld in lib, but
on Windows, we install lass.py in bin. So we set the target's
LIBRARY_OUTPUT_DIRECTORY to match.

[1] https://cmake.org/cmake/help/v3.28/manual/cmake-buildsystem.7.html#library-output-artifacts