Fix package_info for Conan 1
In editable mode, self.package_folder is empty, which broke the
try/except to load LassConfig.py. And in Conan 2, self.build_folder is
empty in editable mode.

PATH is only extended with bindirs in Conan 2, so for test_package to
properly run in Conan 1, we must extend PATH manually.