Making QuadTree ray-intersect(s) more robust
Adding a more challenging intersection test in test_spat_object_trees
where we look for the first two intersections along a ray (by using the
tHit of the first intersection as tMin for the second).

This revealed some troubles with the QuadTree ray-intersect and
ray-intersects. The same treatment as for the ray-find helped here as
well: recomputing tMiddle from the node's center. Even though we're no
longer scaling up the child nodes that caused the runaway of tMiddle.