Return existing module when already injected
And increase refcount of returned module.

Following scenario causes a module to be imported and initialized twice:

  import lass
  del sys.modules["lass"]
  import lass

(Without deleting the module from sys.modules, the second import would
simply return the existing module object, instad of fully importing it.

In debug mode, this results in the following assertion failure:

  lass::python::ModuleDefinition::inject: ASSERT FAILURE: '!isInjected_'

In release mode, this results in the module to be injected again, and at
least causes a memory leak of the old module_ instance.

Without fully implementing isolated modules [1] with multi-phase
initialization and heap-allocated types, the best we can do for now is
to return the existing module_ instance.

PyInit_<modulename>() also is expected to return a new reference to the
created module, so we must increase refcount if we return the module_
object.

[1] https://docs.python.org/3.11/howto/isolating-extensions.html