Add QbvhTree, an 4-ary AABB tree
An AABB tree that stores 4 children per node [1]. Key difference with the
AabbTree is that the bounding boxes of the children are stored in the
parent node, rather than each child node having its own bounding box.
This is better for cache locals, as we need to test all 4 bounding boxes
at once. We use AVX instructions and the tests from [2], to accelerate
the tests. We use the gamma error bounds from previous commit, to
conservatively tests the bounding boxes [3, pp. 220-221].

[1] H. Dammertz, J. Hanika, and A. Keller. Shallow bounding volume
    hierarchies for fast SIMD ray tracing of incoherent rays. In
    Proceedings of the Nineteenth Eurographics conference on Rendering
    2008 (EGSR '08). Eurographics Association, Goslar, DEU, 1225–1233.
    https://doi.org/10.1111/j.1467-8659.2008.01261.x
[2] https://tavianator.com/2022/ray_box_boundary.html
[3] M. Phar, W. Jakob, G. Humphreys, Physically Based Rendering: From
    Theory To Implementation, 3rd ed., San Francisco, CA, USA:
    Morgan Kaufmann, 2017. ISBN 978-0-12-800645-0
    https://www.pbr-book.org/3ed-2018/Shapes/Managing_Rounding_Error#ConservativeRayndashBoundsIntersections